+++ src/app.module.ts
@@ -1,13 +1,20 @@
 import { Logger, Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
-import { NacosModule } from '@wlisfes/chat-web-base-schema/nacos'
+import { createNacosRuntimeOptions, NacosModule } from '@wlisfes/chat-web-base-schema/nacos'
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
 
 @Module({
     imports: [
         ConfigModule.forRoot({ isGlobal: true }),
-        NacosModule.forRoot({ serviceName: 'chat-web-skyline-service', registerPort: 3000 })
+        NacosModule.forRoot(
+            createNacosRuntimeOptions({
+                serviceName: process.env.NACOS_SERVICE_NAME!,
+                registerPort: process.env.PORT,
+                NACOS_SERVER: process.env.NACOS_SERVER,
+                NACOS_NAMESPACE: process.env.NACOS_NAMESPACE
+            })
+        )
     ],
     controllers: [AppController],
     providers: [Logger, AppService]